# Conflicts:
#	docs/conf.py

+++ conflowgen/domain_models/distribution_models/truck_arrival_distribution.py
@@ -1,11 +1,11 @@
-from peewee import FloatField, IntegerField
+from peewee import FloatField
 
 from conflowgen.domain_models.base_model import BaseModel
 
 
 class TruckArrivalDistribution(BaseModel):
     """The truck arrival distribution (both inbound and outbound journeys)"""
-    hour_in_the_week = IntegerField(null=False, primary_key=True, unique=True)
+    hour_in_the_week = FloatField(null=False, primary_key=True, unique=True)
     fraction = FloatField(null=False)
 
     def __repr__(self):
